[overrides] Remove unused visitor and fix path of other
