+++ package.json
@@ -17,7 +17,7 @@
     "workbench:react-native:d8b0997a:check": "yarn workbench:react-native:d8b0997a:check:ios && yarn workbench:react-native:d8b0997a:check:android",
     "workbench:react-native:d8b0997a:cli": "yarn workspace flow2dts cli --cwd ../.. --rootDir ./workbench/react-native/d8b0997a/inputs --overrides ./workbench/react-native/d8b0997a/overrides.ts --hint ./workbench/react-native/d8b0997a/hint/hint.json",
     "workbench:react-native:d8b0997a:cli-with-globs": "yarn workbench:react-native:d8b0997a:cli \"./workbench/react-native/d8b0997a/inputs/{index,Libraries/**/*}.js.flow\" \"!./workbench/react-native/d8b0997a/inputs/Libraries/**/{__flowtests__,__mocks__,__tests__}/**/*.js.flow\" \"!./workbench/react-native/d8b0997a/inputs/Libraries/{Inspector,JSInspector,NewAppScreen,ReactPrivate}/**/*.js.flow\" \"!./workbench/react-native/d8b0997a/inputs/Libraries/Animated/AnimatedWeb.js.flow\" \"!./workbench/react-native/d8b0997a/inputs/Libraries/{Promise,promiseRejectionIsError,promiseRejectionTrackingOptions}.js.flow\" \"!./workbench/react-native/d8b0997a/inputs/Libraries/vendor/core/ErrorUtils.js.flow\"",
-    "workbench:react-native:d8b0997a:ios": "rimraf ./workbench/react-native/d8b0997a/outputs/ios/Libraries && yarn workbench:react-native:d8b0997a:cli-with-globs --platform ios --outDir ./workbench/react-native/d8b0997a/outputs/ios",
+    "workbench:react-native:d8b0997a:ios": "rimraf ./workbench/react-native/d8b0997a/outputs/ios/Libraries && yarn workbench:react-native:d8b0997a:cli-with-globs --platform ios --overrides ./workbench/react-native/d8b0997a/overrides.ios.ts --outDir ./workbench/react-native/d8b0997a/outputs/ios",
     "workbench:react-native:d8b0997a:android": "rimraf ./workbench/react-native/d8b0997a/outputs/android/Libraries && yarn workbench:react-native:d8b0997a:cli-with-globs --platform android --outDir ./workbench/react-native/d8b0997a/outputs/android",
     "workbench:react-native:d8b0997a:check:ios": "yarn tsc --project ./workbench/react-native/d8b0997a/outputs/ios/tsconfig.json --pretty",
     "workbench:react-native:d8b0997a:check:android": "yarn tsc --project ./workbench/react-native/d8b0997a/outputs/android/tsconfig.json --pretty",
@@ -25,7 +25,7 @@
     "workbench:react-native:v0.63.3:check": "yarn workbench:react-native:v0.63.3:check:ios && yarn workbench:react-native:v0.63.3:check:android",
     "workbench:react-native:v0.63.3:cli": "yarn workspace flow2dts cli --cwd ../.. --rootDir ./workbench/react-native/v0.63.3/inputs --overrides ./workbench/react-native/v0.63.3/overrides.ts --hint ./workbench/react-native/v0.63.3/hint/hint.json",
     "workbench:react-native:v0.63.3:cli-with-globs": "yarn workbench:react-native:v0.63.3:cli \"./workbench/react-native/v0.63.3/inputs/{index,Libraries/**/*}.js.flow\" \"!./workbench/react-native/v0.63.3/inputs/Libraries/**/{__flowtests__,__mocks__,__tests__}/**/*.js.flow\" \"!./workbench/react-native/v0.63.3/inputs/Libraries/{Inspector,JSInspector,NewAppScreen,polyfills,ReactPrivate}/**/*.js.flow\" \"!./workbench/react-native/v0.63.3/inputs/Libraries/Animated/AnimatedWeb.js.flow\" \"!./workbench/react-native/v0.63.3/inputs/Libraries/{Promise,promiseRejectionIsError,promiseRejectionTrackingOptions}.js.flow\" \"!./workbench/react-native/v0.63.3/inputs/Libraries/vendor/core/ErrorUtils.js.flow\"",
-    "workbench:react-native:v0.63.3:ios": "rimraf ./workbench/react-native/v0.63.3/outputs/ios/Libraries && yarn workbench:react-native:v0.63.3:cli-with-globs --platform ios --outDir ./workbench/react-native/v0.63.3/outputs/ios",
+    "workbench:react-native:v0.63.3:ios": "rimraf ./workbench/react-native/v0.63.3/outputs/ios/Libraries && yarn workbench:react-native:v0.63.3:cli-with-globs --platform ios --overrides ./workbench/react-native/d8b0997a/overrides.ios.ts --outDir ./workbench/react-native/v0.63.3/outputs/ios",
     "workbench:react-native:v0.63.3:android": "rimraf ./workbench/react-native/v0.63.3/outputs/android/Libraries && yarn workbench:react-native:v0.63.3:cli-with-globs --platform android --outDir ./workbench/react-native/v0.63.3/outputs/android",
     "workbench:react-native:v0.63.3:check:ios": "yarn tsc --project ./workbench/react-native/v0.63.3/outputs/ios/tsconfig.json --pretty",
     "workbench:react-native:v0.63.3:check:android": "yarn tsc --project ./workbench/react-native/v0.63.3/outputs/android/tsconfig.json --pretty",
