+++ workbench/react-native/v0.63.3/overrides.ts
@@ -288,22 +288,6 @@ const visitors: OverridesVisitor[] = [
       },
     },
   ],
-  [
-    "Libraries/promiseRejectionTrackingOptions.d.ts",
-    {
-      DeclareVariable: {
-        exit(path) {
-          if (path.node.id.name === "rejectionTrackingOptions") {
-            const replacementDeclaration = ast`
-            declare var rejectionTrackingOptions: Parameters<enable>[0]
-          ` as t.VariableDeclaration
-            path.replaceWith(replacementDeclaration)
-            path.skip()
-          }
-        },
-      },
-    },
-  ],
   // TODO: This should be fixed upstream in RN. No duplicate prop entries should exist.
   [
     "Libraries/Text/TextProps.d.ts",
@@ -320,23 +304,6 @@ const visitors: OverridesVisitor[] = [
             if (state.seen_adjustsFontSizeToFit_times === 2) {
               path.remove()
             }
-          }
-        },
-      },
-    },
-  ],
-  // TODO: This should be fixed upstream in RN. This seems like simply broken upstream code.
-  [
-    "Libraries/Components/SegmentedControlIOS/SegmentedControlIOS.ios.d.ts",
-    {
-      TSTypeReference: {
-        exit(path) {
-          if (t.isIdentifier(path.node.typeName) && path.node.typeName.name === "NativeSegmentedControlIOS") {
-            const typeAlias = ast`
-            type Replacement = typeof import("./RCTSegmentedControlNativeComponent").default
-          ` as t.TSTypeAliasDeclaration
-            path.replaceWith(typeAlias.typeAnnotation)
-            path.skip()
           }
         },
       },
@@ -395,28 +362,6 @@ const visitors: OverridesVisitor[] = [
             t.assertTSTypeAnnotation(renderMethod.returnType)
             t.assertTSVoidKeyword(renderMethod.returnType.typeAnnotation)
             renderMethod.returnType.typeAnnotation = t.tsNeverKeyword()
-          }
-        },
-      },
-    },
-  ],
-  [
-    "Libraries/Components/AccessibilityInfo/AccessibilityInfo.ios.d.ts",
-    {
-      ImportDeclaration: {
-        exit(path, state: { promiseIdentifierName?: string }) {
-          if (path.node.source.value === "../../Promise") {
-            const specifier = path.node.specifiers[0]
-            t.assertImportDefaultSpecifier(specifier)
-            state.promiseIdentifierName = specifier.local.name
-            path.remove()
-          }
-        },
-      },
-      TSTypeReference: {
-        exit(path, state: { promiseIdentifierName: string }) {
-          if (t.isIdentifier(path.node.typeName) && path.node.typeName.name === state.promiseIdentifierName) {
-            path.node.typeName = t.identifier("Promise")
           }
         },
       },
